test working wiht User
