+++ src/models/churchUser.ts
@@ -4,7 +4,7 @@ import { Church } from "./church";
 
 export class ChurchUser extends Model<InferAttributes<ChurchUser>, InferCreationAttributes<ChurchUser>>{
     declare userId: number;
-    declare churchId: number;
+    
     declare email: string;
     declare password: string;
     
@@ -18,10 +18,7 @@ export function ChurchUserFactory( sequelize: Sequelize ) {
             primaryKey: true,
             allowNull: false
         },
-        churchId:{
-            type: DataTypes.INTEGER,
-            allowNull: false
-        },
+        
         
         email: {
             type: DataTypes.STRING,
@@ -43,7 +40,7 @@ export function ChurchUserFactory( sequelize: Sequelize ) {
 
 
 
-export function AssociateUserChurch(){
-    ChurchUser.hasMany(Church, { foreignKey: 'userId' });
-    Church.belongsTo(ChurchUser, { foreignKey: 'userId' });
-}
+// export function AssociateUserChurch(){
+//     ChurchUser.hasMany(Church, { foreignKey: 'userId' });
+//     Church.belongsTo(ChurchUser, { foreignKey: 'userId' });
+// }
