+++ src/models/index.ts
@@ -1,11 +1,12 @@
 import { Sequelize } from "sequelize";
-import { ChurchUserFactory, AssociateUserChurch } from "./churchUser";
-import { ChurchFactory } from "./church";
+import { ChurchUserFactory } from "./churchUser";
+import { ChurchFactory, AssociateUserChurch } from "./church";
 import { AssociateChurchEvent, EventFactory } from "./event";
-
+//  
 const dbName = 'ChurchDB';
 const username = 'root';
-const password = 'password1!';
+const password = 'Password1!';
+// const password = 'password1!';
 // const password = '0624';
